rename SetInjectToAddTagRector to generic MethodCallToAnotherMethodCallWithArgumentsRector
