+++ packages/Nette/tests/Rector/DI/MethodCallToAnotherMethodCallWithArgumentsRector/Correct/correct.php.inc
