+++ packages/Nette/tests/Rector/DI/MethodCallToAnotherMethodCallWithArgumentsRector/Source/NetteServiceDefinition.php
