+++ packages/Nette/tests/Rector/DI/MethodCallToAnotherMethodCallWithArgumentsRector/Wrong/wrong.php.inc
