+++ packages/Nette/tests/Rector/DI/MethodCallToAnotherMethodCallWithArgumentsRector/config.yml
@@ -1,3 +1,3 @@
 services:
-    Rector\Nette\Rector\DI\SetInjectToAddTagRector:
+    Rector\Nette\Rector\DI\MethodCallToAnotherMethodCallWithArgumentsRector:
         $serviceDefinitionClass: 'Rector\Nette\Tests\Rector\DI\SetInjectToAddTagRector\Source\NetteServiceDefinition'
